minor bug fix

+++ audioSegmentation.py
@@ -638,7 +638,7 @@ def silenceRemoval(x, Fs, stWin, stStep, smoothWindow=0.5, Weight=0.5, plot=Fals
 
     # Step 4A: detect onset frame indices:
     ProbOnsetSorted = numpy.sort(ProbOnset)                        # find probability Threshold as a weighted average of top 10% and lower 10% of the values
-    Nt = ProbOnsetSorted.shape[0] / 10
+    Nt = int(ProbOnsetSorted.shape[0] / 10)
     T = (numpy.mean((1 - Weight) * ProbOnsetSorted[0:Nt]) + Weight * numpy.mean(ProbOnsetSorted[-Nt::]))
 
     MaxIdx = numpy.where(ProbOnset > T)[0]                         # get the indices of the frames that satisfy the thresholding
